pyproject.toml
[project]
name = "voice-agent-hub"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "sqlalchemy>=2.0",
  "pyjwt>=2.9",
  "cryptography>=43",
  "livekit-api>=0.7",
  "livekit>=0.17",
  "httpx>=0.27",
  "python-multipart>=0.0.12",
  "aiosqlite>=0.20",
]

[project.optional-dependencies]
test = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "httpx>=0.27",
  "ruff>=0.9",
  "openai>=2.29.0",
  "psutil>=5.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "--ignore=tests/integration --ignore=tests/smoke_test.py"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501", "B008"]
